requirements.txt
action-msgs==1.2.1
action-tutorials-interfaces==0.20.5
action-tutorials-py==0.20.5
actionlib-msgs==4.8.0
ament-cmake-test==1.3.11
ament-copyright==0.12.12
ament-cppcheck==0.12.12
ament-cpplint==0.12.12
ament-flake8==0.12.12
ament-index-python==1.4.0
ament-lint==0.12.12
ament-lint-cmake==0.12.12
ament-package==0.14.0
ament-pep257==0.12.12
ament-uncrustify==0.12.12
ament-xmllint==0.12.12
angles==1.15.0
apriltag-msgs==2.0.1
aruco-algorithm==0.0.0
aurco-algorithm==0.0.0
builtin-interfaces==1.2.1
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
common-msgs==0.0.0
composition-interfaces==1.2.1
cv-bridge==3.2.1
demo-nodes-py==0.20.5
diagnostic-msgs==4.8.0
diagnostic-updater==4.0.2
domain-coordinator==0.10.0
example-interfaces==0.9.3
examples-rclpy-executors==0.15.3
examples-rclpy-minimal-action-client==0.15.3
examples-rclpy-minimal-action-server==0.15.3
examples-rclpy-minimal-client==0.15.3
examples-rclpy-minimal-publisher==0.15.3
examples-rclpy-minimal-service==0.15.3
examples-rclpy-minimal-subscriber==0.15.3
filelock==3.20.0
fsspec==2025.12.0
ftfy==6.3.1
geometry-msgs==4.8.0
image-geometry==3.2.1
interactive-markers==2.3.2
io-module==0.0.0
irhx==0.0.0
Jinja2==3.1.6
laser-geometry==2.4.0
launch==1.0.8
launch-pytest==1.0.8
launch-ros==0.19.9
launch-testing==1.0.8
launch-testing-ros==0.19.9
launch-xml==1.0.8
launch-yaml==1.0.8
lifecycle-msgs==1.2.1
logging-demo==0.20.5
map-msgs==2.1.0
MarkupSafe==2.1.5
message-filters==4.3.6
model-module==0.0.0
motion-control-module==0.0.0
motor-control==0.0.0
motor-control-module==0.0.0
mpmath==1.3.0
nav-msgs==4.8.0
networkx==3.4.2
numpy==2.2.6
object-frame-compensation-package==0.0.0
opencv-python==4.13.0.90
osrf-pycommon==2.1.6
packaging==26.0
pcl-msgs==1.0.0
pendulum-msgs==0.20.5
pillow==12.0.0
python-qt-binding==1.1.2
qt-dotgraph==2.2.4
qt-gui==2.2.4
qt-gui-cpp==2.2.4
qt-gui-py-common==2.2.4
quality-of-service-demo-py==0.20.5
rcl-interfaces==1.2.1
rclpy==3.3.16
rcutils==5.1.6
realsense==0.0.0
realsense-camera-driver==0.0.1
regex==2026.1.15
resource-retriever==3.1.3
rmw-dds-common==1.6.0
robot-controller-ros2==1.0.0
robot-controller-ros2-interfaces==0.0.0
ros2action==0.18.12
ros2bag==0.15.14
ros2cli==0.18.12
ros2component==0.18.12
ros2doctor==0.18.12
ros2interface==0.18.12
ros2launch==0.19.9
ros2lifecycle==0.18.12
ros2multicast==0.18.12
ros2node==0.18.12
ros2param==0.18.12
ros2pkg==0.18.12
ros2run==0.18.12
ros2service==0.18.12
ros2topic==0.18.12
rosbag2-interfaces==0.15.14
rosbag2-py==0.15.14
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.6
rosidl-cli==3.1.6
rosidl-cmake==3.1.6
rosidl-generator-c==3.1.6
rosidl-generator-cpp==3.1.6
rosidl-generator-py==0.14.4
rosidl-parser==3.1.6
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.6
rosidl-typesupport-introspection-cpp==3.1.6
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-msg==1.2.0
rqt-plot==1.1.4
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
sensor-msgs==4.9.0
sensor-msgs-py==4.8.0
shape-msgs==4.8.0
show-image-node==0.0.1
sros2==0.10.6
state-machine-module==0.0.0
statistics-msgs==1.2.1
std-msgs==4.8.0
std-srvs==4.8.0
stereo-msgs==4.8.0
sympy==1.14.0
task-module==0.0.0
teleop-twist-keyboard==2.4.0
tf2-geometry-msgs==0.25.12
tf2-kdl==0.25.12
tf2-msgs==0.25.12
tf2-py==0.25.12
tf2-ros-py==0.25.12
tf2-tools==0.25.12
topic-monitor==0.20.5
torch==2.10.0+cpu
torchvision==0.25.0+cpu
tqdm==4.67.1
trajectory-msgs==4.8.0
turtlesim==1.4.2
typing_extensions==4.15.0
ui-app==0.0.0
ui-module==0.0.0
unique-identifier-msgs==2.2.1
uros-interface==0.0.0
vision-compensation==0.0.1
vision-dector==0.0.0
vision-module==0.0.0
visualization-msgs==4.8.0
wcwidth==0.5.0
xacro==2.0.13
